# Optimize ParamResolver.value_of