--- a/cirq-core/cirq/study/resolver_test.py
+++ b/cirq-core/cirq/study/resolver_test.py
@@ -53,10 +53,10 @@ def test_value_of_transformed_types(val, resolved):
 
 @pytest.mark.parametrize('val,resolved', [(sympy.I, 1j)])
 def test_value_of_substituted_types(val, resolved):
-    _assert_consistent_resolution(val, resolved, True)
+    _assert_consistent_resolution(val, resolved)
 
 
-def _assert_consistent_resolution(v, resolved, subs_called=False):
+def _assert_consistent_resolution(v, resolved):
     """Asserts that parameter resolution works consistently.
 
     The ParamResolver.value_of method can resolve any Sympy expression -
@@ -70,7 +70,7 @@ def _assert_consistent_resolution(v, resolved, subs_called=False):
     Args:
         v: the value to resolve
         resolved: the expected resolution result
-        subs_called: if True, it is expected that the slow subs method is called
+
     Raises:
         AssertionError in case resolution assertion fail.
     """
@@ -93,9 +93,7 @@ def _assert_consistent_resolution(v, resolved, subs_called=False):
     # symbol based resolution
     s = SubsAwareSymbol('a')
     assert r.value_of(s) == resolved, f"expected {resolved}, got {r.value_of(s)}"
-    assert (
-        subs_called == s.called
-    ), f"For pass-through type {type(v)} sympy.subs shouldn't have been called."
+    assert not s.called, f"For pass-through type {type(v)} sympy.subs shouldn't have been called."
     assert isinstance(
         r.value_of(s), type(resolved)
     ), f"expected {type(resolved)} got {type(r.value_of(s))}"
@@ -243,15 +241,18 @@ def test_custom_resolved_value():
 
 
 def test_custom_value_not_implemented():
-    class Bar:
+    class BarImplicit:
+        pass
+
+    class BarExplicit:
         def _resolved_value_(self):
             return NotImplemented
 
-    b = sympy.Symbol('b')
-    bar = Bar()
-    r = cirq.ParamResolver({b: bar})
-    with pytest.raises(sympy.SympifyError):
-        _ = r.value_of(b)
+    for cls in [BarImplicit, BarExplicit]:
+        b = sympy.Symbol('b')
+        bar = cls()
+        r = cirq.ParamResolver({b: bar})
+        assert r.value_of(b) == b
 
 
 def test_compose():