@andbe91 2023-11-14
Any chance of getting this into master this week? Would be great for accelerating some internal efforts :)